+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nifti data type field validation
 - Unit field validation
 - Pydantic-based validation/data sanitising
-- Generic ground truth loader filter.
+- Generic ground truth loader filter with input validation.
 - General maths utility function and filter.
 - Filter that assigns values to regions specified in a segmentation mask.
 - Filter validation utility function.